Extract processing information modal content into it's own file
