+++ src/containers/Experiment/ProcessingInformation/ProcessingInformation.scss
@@ -1,5 +1,5 @@
-@import '../../common/styles/variables';
-@import '../../common/styles/mixins';
+@import '../../../common/styles/variables';
+@import '../../../common/styles/mixins';
 
 .processing-info-modal {
   padding: 56px;
